+++ enterprise/kernel/LICENSES.txt
