+++ enterprise/kernel/src/main/java/org/neo4j/graphdb/EnterpriseGraphDatabase.java
@@ -24,7 +24,7 @@ import java.util.Map;
 
 import org.neo4j.graphdb.factory.GraphDatabaseSettings;
 import org.neo4j.kernel.extension.KernelExtensionFactory;
-import org.neo4j.kernel.ha.factory.EnterpriseFacadeFactory;
+import org.neo4j.kernel.impl.enterprise.EnterpriseFacadeFactory;
 import org.neo4j.kernel.impl.factory.GraphDatabaseFacade;
 import org.neo4j.kernel.impl.factory.GraphDatabaseFacadeFactory;
 import org.neo4j.kernel.monitoring.Monitors;
