+++ enterprise/kernel/src/main/java/org/neo4j/kernel/impl/enterprise/EnterpriseConstraintSemantics.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.kernel.ha.factory;
+package org.neo4j.kernel.impl.enterprise;
 
 import java.util.Iterator;
 
@@ -45,7 +45,7 @@ import org.neo4j.kernel.impl.store.record.RelationshipPropertyExistenceConstrain
 import static org.neo4j.kernel.impl.store.record.NodePropertyExistenceConstraintRule.nodePropertyExistenceConstraintRule;
 import static org.neo4j.kernel.impl.store.record.RelationshipPropertyExistenceConstraintRule.relPropertyExistenceConstraintRule;
 
-class EnterpriseConstraintSemantics extends StandardConstraintSemantics
+public class EnterpriseConstraintSemantics extends StandardConstraintSemantics
 {
     @Override
     protected PropertyConstraint readNonStandardConstraint( PropertyConstraintRule rule )
