+++ enterprise/kernel/src/test/java/org/neo4j/SchemaHelper.java
