+++ enterprise/kernel/src/test/java/org/neo4j/shell/TestPECListing.java
@@ -28,7 +28,6 @@ import org.neo4j.graphdb.Label;
 import org.neo4j.graphdb.RelationshipType;
 import org.neo4j.kernel.GraphDatabaseAPI;
 import org.neo4j.test.TestEnterpriseGraphDatabaseFactory;
-import org.neo4j.test.TestGraphDatabaseFactory;
 
 import static org.neo4j.graphdb.DynamicLabel.label;
